rename chains to chain
