+++ plugin/evm/test_tx.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/ava-labs/avalanchego/utils"
 
-	"github.com/ava-labs/avalanchego/chains/atomic"
+	"github.com/ava-labs/avalanchego/chain/atomic"
 	"github.com/ava-labs/avalanchego/codec"
 	"github.com/ava-labs/avalanchego/codec/linearcodec"
 	"github.com/ava-labs/avalanchego/ids"
